+++ docs/changelog.md
@@ -89,6 +89,45 @@
 
 ## 更新说明详情
 
+1.18.25 fix: fix qdii_e_comm_jsl interface
+
+	1. 修复 qdii_e_comm_jsl 接口
+	2. 修复 stock_rank_xstp_ths 接口
+
+1.18.24 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+
+1.18.23 fix: fix fund_fee_em interface
+
+	1. 修复 fund_fee_em 接口
+	2. 修复 fund_etf_category_ths 接口
+
+1.18.22 fix: fix bond_gb_zh_sina interface
+
+	1. 修复 bond_gb_zh_sina 接口
+	2. 修复 bond_gb_us_sina 接口
+
+1.18.21 fix: fix get_receipt interface
+
+	1. 修复 get_receipt 接口
+
+1.18.20 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+
+1.18.19 fix: fix GitHub action
+
+	1. 修复配置文件
+
+1.18.18 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+
+1.18.17 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+
 1.18.16 fix: fix news_economic_baidu interface
 
 	1. 修复 news_economic_baidu 接口
@@ -598,7 +637,7 @@
 	1. 修复 stock_info_sz_delist 接口
 
 1.16.98 fix: fix fund_open_fund_info_em interface
-	
+
 	1. 修复 fund_open_fund_info_em 接口
 
 1.16.97 fix: fix fund_open_fund_info_em interface
@@ -606,7 +645,7 @@
 	1. 修复 fund_open_fund_info_em 接口
 
 1.16.96 fix: fix stock_individual_spot_xq interface
-	
+
 	1. 修复 stock_individual_spot_xq 接口
 	2. 修复 stock_cyq_em 接口
 
@@ -6093,6 +6132,24 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.18.25 fix: fix qdii_e_comm_jsl interface
+
+1.18.24 fix: fix stock_individual_spot_xq interface
+
+1.18.23 fix: fix fund_fee_em interface
+
+1.18.22 fix: fix bond_gb_zh_sina interface
+
+1.18.21 fix: fix get_receipt interface
+
+1.18.20 fix: fix stock_individual_spot_xq interface
+
+1.18.19 fix: fix GitHub action
+
+1.18.18 fix: fix stock_individual_spot_xq interface
+
+1.18.17 fix: fix stock_individual_spot_xq interface
 
 1.18.16 fix: fix news_economic_baidu interface